first attempt to render as sphere or point depending on size of sphere.
Basically, spheres won't be rendered at all if you zoom out, because
they become less than a pixel in size. GL_POINTS are rendered regardless
of zoom. So I'm attempting to calculate the pixel diameter of the sphere
before rendering. If it would be too small, I just draw a point instead.
The current approach to calculating the diameter works inconsistently,
and will need to be fixed